Revert "feat: Rabbitmq heavy task processing isolation"
