--- src/core/utils/stream-to-buffer.ts
@@ -1,8 +1,0 @@
-export async function streamToBuffer(stream: ReadableStream): Promise<Buffer> {
-  const chunks: Buffer[] = [];
-  for await (const chunk of stream) {
-    chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-  }
-
-  return Buffer.concat(chunks);
-}
